Expand GP Shop with event items and improve daily missions
GP Shop Enhancements:
- Add Event & Collab Items category (Category 15) with free promotional items
- Include Overlord collaboration items (Ainz staff, Albedo dress, tapestries)
- Include Berserk collaboration items (Guts greatsword, equipment)
- Include Olfring promotional weapons (dagger, cane, bow)
- Add NPC outfit costumes (Fabio, Klaus, Iris, Lise, Diamantes, Vanessa, Mayleaf)
- Reorganize categories: Gold & Currency, Dungeon Tickets, Training Sets, Bundles

Daily Mission Improvements:
- Add DailyMissionEventListGetHandler for event missions
- Fix CDataDailyMissionInfo structure and serialization
- Add CDataDailyMissionSecondEntry for additional mission data
- Add comprehensive tests for daily mission serialization

Other Fixes:
- Update DispelGetDispelItemSettingsHandler
- Fix InstanceEnemyKillHandler
- Expand EventCodes.json with more event code entries
- Add new ShopType entries

+++ Arrowgene.Ddon.Shared/Entity/Structure/CDataDailyMissionReward.cs
@@ -20,7 +20,7 @@ namespace Arrowgene.Ddon.Shared.Entity.Structure
             {
                 WriteUInt32(buffer, obj.ItemId);
                 WriteUInt32(buffer, obj.Num);
-                WriteByte(buffer, (byte)obj.WalletType);
+                WriteUInt32(buffer, (uint)obj.WalletType);
                 WriteUInt32(buffer, obj.WalletAmount);
             }
 
@@ -29,7 +29,7 @@ namespace Arrowgene.Ddon.Shared.Entity.Structure
                 CDataDailyMissionReward obj = new CDataDailyMissionReward();
                 obj.ItemId = ReadUInt32(buffer);
                 obj.Num = ReadUInt32(buffer);
-                obj.WalletType = (WalletType)ReadByte(buffer);
+                obj.WalletType = (WalletType)ReadUInt32(buffer);
                 obj.WalletAmount = ReadUInt32(buffer);
                 return obj;
             }
